Adding systemwide installation option

+++ install.sh
@@ -1,14 +1,25 @@
 #!/usr/bin/env bash
 
-release="v2.3.0-release"
+release="v2.5.0-release"
 
-#Installing on local folder on HOME
-dest="$HOME/.local/bin"
-share="$HOME/.local/share/epicsmng"
-configdir="$HOME/.config/epicsmng"
-completion_dir="$HOME/.bash_completion.d"
+#Path locations
+dest="/usr/local/bin"
+share="/usr/local/share/epicsmng"
+src="$share/src"
+configdir="/etc/epicsmng"
+completion_dir="/etc/bash_completion.d"
 settingsdir="$configdir/settings"
 patchesdir="$configdir/patches"
+
+if [[ $UID != 0 ]]; then
+    #Path locations
+    dest="$HOME/.local/bin"
+    share="$HOME/.local/share/epicsmng"
+    configdir="$HOME/.config/epicsmng"
+    completion_dir="$HOME/.bash_completion.d"
+    settingsdir="$configdir/settings"
+    patchesdir="$configdir/patches"
+fi
 
 # cd to repository folder
 cd "$(dirname "$0")" || exit 1
@@ -24,24 +35,38 @@ if ! sed -e "s/_VERSION_/${version}/" ./epicsmng > epicsmng_tmp; then
     exit 1
 fi
 
+if ! sed -i -e "s/_SHARE_DIR_/${share//\//\\/}/" ./epicsmng_tmp; then
+    echo "Installation failed"
+    exit 1
+fi
+
+if ! sed -i -e "s/_CONF_DIR_/${configdir//\//\\/}/" ./epicsmng_tmp; then
+    echo "Installation failed"
+    exit 1
+fi
+
 if ! install -m 755 ./epicsmng_tmp "$dest/epicsmng"; then
     echo "Installation failed"
     exit 1
 fi
 rm epicsmng_tmp
 
-#Update PATH if necessary
-if ! echo "$PATH" | grep -q "$HOME/.local/bin"; then
-    echo "export PATH=\$PATH:$dest # Added by epicsmng" >> "$HOME"/.bashrc
-    echo "PATH updated: open a new shell to start using epicsmng"
+if [[ $UID != 0 ]]; then
+    #Update PATH if necessary
+    if ! echo "$PATH" | grep -q "$HOME/.local/bin"; then
+        echo "export PATH=\$PATH:$dest # Added by epicsmng" >> "$HOME"/.bashrc
+        echo "PATH updated: open a new shell to start using epicsmng"
+    fi
+
+    #Install the bash completion include
+    install -d "$completion_dir"
+    if ! grep -qs "Installed by epicsmng" "$HOME/.bash_completion"; then
+        [ -f "$HOME/.bash_completion" ] && mv "$HOME/.bash_completion"  "$completion_dir/original.bash"
+        cp "$utils/completion_include.bash" "$HOME/.bash_completion"
+    fi
 fi
 
 #Install the bash completion
-install -d "$completion_dir"
-if ! grep -qs "Installed by epicsmng" "$HOME/.bash_completion"; then
-    [ -f "$HOME/.bash_completion" ] && mv "$HOME/.bash_completion"  "$completion_dir/original.bash"
-    cp "$utils/completion_include.bash" "$HOME/.bash_completion"
-fi
 install -m 644 "$utils/epicsmng-completion.bash" "$completion_dir"
 
 #Remove existing sources to avoid conflicts
@@ -56,12 +81,22 @@ if ! install -d "$patchesdir"; then
     echo "WARNING: cannot create patches directory $patchesdir"
 fi
 
+if ! install -m 777 -d "$share"; then
+    echo "ERROR: cannot create source directory $share"
+    exit 1
+fi
+
+if ! install -m 777 -d "$src"; then
+    echo "ERROR: cannot create source directory $src"
+    exit 1
+fi
+
 #Install default config overwriting existing one
-cp "$utils/default.settings" "$configdir"
+install -m 664 "$utils/default.settings" "$configdir"
 
 #if configdir empty, populate it with example user config
 if [ -z "$(ls -A "$settingsdir")" ]; then
-    cp "$utils/user.settings" "$settingsdir"
+    install -m 664 "$utils/user.settings" "$settingsdir"
 fi
 
 echo "Done!"
